Backend/requirements.txt
alembic==1.13.1
asttokens==2.4.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
distlib==0.3.7
executing==2.0.1
filelock==3.13.1
Flask==3.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.6
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
Mako==1.3.1
MarkupSafe==2.1.4
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pipenv==2023.11.15
platformdirs==4.1.0
pluggy==1.3.0
prompt-toolkit==3.0.41
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
pytest==7.4.3
python-dotenv==1.0.1
requests==2.31.0
setuptools==69.0.3
six==1.16.0
SQLAlchemy==2.0.25
SQLAlchemy-serializer==1.4.1
stack-data==0.6.3
traitlets==5.14.0
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.0
wcwidth==0.2.12
Werkzeug==3.0.1
wheel==0.42.0
